Revert "changed code"
This reverts commit d115202fc27268567e2e61c63dc0f260f4ce8c0c.

+++ Code/DC2019_MASTER/DC2019_MASTER.ino
@@ -38,7 +38,7 @@ String readBTline(){
 
 void distance_detect(){
   long duration = 0;
-  long distance;
+  int distance;
   for (int i = 0; i<3; i++){
     // Clears the trigPin
     digitalWrite(trigPin, LOW);
